+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+* Added `FullVersion::parse` convenience method
+
 ## [0.3.0]
 
 ### Added